Port to musl libc
Closes #72 and uses some patches from #193 for better portability
between UNIXes.

Also this patch replaces so many NULLs with 0 or '\0' because musl
headers redefines it and anyway it is better practice.

Moreover, due to very "cool" code from Valve, some modules comes with
disabled fortify which distributes with build-base in Alpine
(fortify is something like compile-time checking for memory bounds)